@@ -25,26 +25,26 @@
     "@angular/cdk": "^11.0.3",
     "@ant-design/icons-angular": "^11.0.1",
     "date-fns": "^2.16.1",
-    "ngx-quicklink": "^0.2.6",
+    "ngx-hover-preload": "0.0.3",
     "resize-observer-polyfill": "^1.5.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.5",
-    "@angular-devkit/core": "^11.0.5",
-    "@angular-devkit/schematics": "^11.0.5",
-    "@angular/animations": "^11.0.5",
-    "@angular/cli": "~11.0.5",
-    "@angular/common": "^11.0.5",
-    "@angular/compiler": "^11.0.5",
-    "@angular/compiler-cli": "~11.0.5",
-    "@angular/core": "^11.0.5",
-    "@angular/forms": "^11.0.5",
-    "@angular/language-service": "^11.0.5",
-    "@angular/platform-browser": "^11.0.5",
-    "@angular/platform-browser-dynamic": "^11.0.5",
-    "@angular/platform-server": "^11.0.5",
-    "@angular/router": "^11.0.5",
-    "@angular/service-worker": "^11.0.5",
+    "@angular-devkit/build-angular": "~0.1101.0",
+    "@angular-devkit/core": "^11.1.0",
+    "@angular-devkit/schematics": "^11.1.0",
+    "@angular/animations": "^11.1.0",
+    "@angular/cli": "~11.1.0",
+    "@angular/common": "^11.1.0",
+    "@angular/compiler": "^11.1.0",
+    "@angular/compiler-cli": "~11.1.0",
+    "@angular/core": "^11.1.0",
+    "@angular/forms": "^11.1.0",
+    "@angular/language-service": "^11.1.0",
+    "@angular/platform-browser": "^11.1.0",
+    "@angular/platform-browser-dynamic": "^11.1.0",
+    "@angular/platform-server": "^11.1.0",
+    "@angular/router": "^11.1.0",
+    "@angular/service-worker": "^11.1.0",
     "@ant-design/dark-theme": "^2.0.2",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-angular": "^11.0.0",
@@ -91,7 +91,7 @@
     "karma-junit-reporter": "^2.0.1",
     "karma-spec-reporter": "0.0.32",
     "karma-viewport": "^1.0.7",
-    "less": "^3.10.3",
+    "less": "^4.1.0",
     "less-plugin-clean-css": "^1.5.1",
     "less-plugin-npm-import": "^2.1.0",
     "less-vars-to-js": "^1.3.0",
@@ -100,10 +100,11 @@
     "marked": "^1.2.7",
     "minimist": "^1.2.5",
     "monaco-editor": "^0.21.2",
-    "ng-packagr": "^11.0.3",
+    "ng-packagr": "^11.1.2",
     "ngx-color": "^6.2.0",
     "node-prismjs": "^0.1.2",
     "parse5": "^6.0.1",
+    "postcss": "^8.2.4",
     "prettier": "^2.2.1",
     "prismjs": "^1.22.0",
     "process": "^0.11.10",
@@ -119,7 +120,7 @@
     "ts-node": "~9.1.1",
     "tslib": "^2.0.3",
     "tslint": "~6.1.3",
-    "typescript": "~4.0.2",
+    "typescript": "~4.1.2",
     "yaml-front-matter": "^4.1.1",
     "zone.js": "~0.11.3"
   },
