@@ -23,9 +23,9 @@ import { Component } from '@angular/core';
 
     <button nz-button nzType="primary" (click)="showModalMiddle()">Vertically centered modal dialog</button>
     <nz-modal
-      nzWrapClassName="vertical-center-modal"
       [(nzVisible)]="isVisibleMiddle"
       nzTitle="Vertically centered modal dialog"
+      nzCentered
       (nzOnCancel)="handleCancelMiddle()"
       (nzOnOk)="handleOkMiddle()"
     >
@@ -35,20 +35,7 @@ import { Component } from '@angular/core';
         <p>some contents...</p>
       </ng-container>
     </nz-modal>
-  `,
-  styles: [
-    `
-      ::ng-deep .vertical-center-modal {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-      }
-
-      ::ng-deep .vertical-center-modal .ant-modal {
-        top: 0;
-      }
-    `
-  ]
+  `
 })
 export class NzDemoModalPositionComponent {
   isVisibleTop = false;
