@@ -3,9 +3,9 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-import { Observable, Subscription } from 'rxjs';
-
+import { TemplateRef } from '@angular/core';
 import { IndexableObject, NzSafeAny } from 'ng-zorro-antd/core/types';
+import { Observable, Subscription } from 'rxjs';
 
 /** Status */
 export type UploadFileStatus = 'error' | 'success' | 'done' | 'uploading' | 'removed';
@@ -89,3 +89,5 @@ export interface NzUploadXHRArgs {
   onSuccess?(ret: NzSafeAny, file: NzUploadFile, xhr: NzSafeAny): void;
   onError?(err: NzSafeAny, file: NzUploadFile): void;
 }
+
+export type NzIconRenderTemplate = TemplateRef<{ $implicit: NzUploadFile }>;
