@@ -5,4 +5,4 @@
 
 import { Version } from '@angular/core';
 
-export const VERSION = new Version('11.0.2');
+export const VERSION = new Version('11.1.0');
