@@ -21,7 +21,7 @@ import { NzMessageModule } from 'ng-zorro-antd/message';
 import { NzPopoverModule } from 'ng-zorro-antd/popover';
 import { NzSelectModule } from 'ng-zorro-antd/select';
 import { ColorSketchModule } from 'ngx-color/sketch';
-import { QuicklinkModule, QuicklinkStrategy } from 'ngx-quicklink';
+import { HoverPreloadModule, HoverPreloadStrategy } from 'ngx-hover-preload';
 
 import { environment } from '../environments/environment';
 import { DEMOComponent } from './_demo/demo.component';
@@ -62,11 +62,11 @@ const icons: IconDefinition[] = [LeftOutline, RightOutline, EditOutline];
     FooterModule,
     NzContributorsListModule,
     FixedWidgetsModule,
-    QuicklinkModule,
+    HoverPreloadModule,
     RouterModule.forRoot(
       routes,
       environment.production
-        ? { preloadingStrategy: QuicklinkStrategy, scrollPositionRestoration: 'enabled', initialNavigation: 'enabledBlocking' }
+        ? { preloadingStrategy: HoverPreloadStrategy, scrollPositionRestoration: 'enabled', initialNavigation: 'enabledBlocking' }
         : {}
     ),
     ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production && !environment.preProduction })
